+++ src/midi.rs
@@ -11,8 +11,8 @@ pub struct Midi {
 impl Midi {
     pub fn new<F, G>(press_callback: F, release_callback: G) -> Midi
     where
-        F: Fn(u32) + std::clone::Clone + std::marker::Send + 'static,
-        G: Fn(u32) + std::clone::Clone + std::marker::Send + 'static,
+        F: Fn(u8) + std::clone::Clone + std::marker::Send + 'static,
+        G: Fn(u8) + std::clone::Clone + std::marker::Send + 'static,
     {
         let callback2 = move |_stamp: u64, message: &[u8], _: &mut ()| {
             let event = LiveEvent::parse(message).unwrap();
@@ -21,15 +21,15 @@ impl Midi {
                 LiveEvent::Midi { channel, message } => match message {
                     MidiMessage::NoteOn { key, vel } if vel == 0 => {
                         println!("release note {} on channel {} vel {}", key, channel, vel);
-                        release_callback(key.as_int() as u32);
+                        release_callback(key.as_int());
                     }
                     MidiMessage::NoteOn { key, vel } => {
                         println!("press note {} on channel {} vel {}", key, channel, vel);
-                        press_callback(key.as_int() as u32);
+                        press_callback(key.as_int());
                     }
                     MidiMessage::NoteOff { key, vel } => {
                         println!("release note {} on channel {} vel {}", key, channel, vel);
-                        release_callback(key.as_int() as u32);
+                        release_callback(key.as_int());
                     }
                     _ => {}
                 },
